pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deep-researcher"
version = "0.1.0"
description = "Autonomous adversarial C4 architecture harness for BMAD projects"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.1.0",
    "pydantic>=2.0",
    "typer>=0.12",
    "gitpython>=3.1",
    "rich>=13.0",
    "openai>=1.0",
]

[project.scripts]
adversarial-architect = "deep_researcher.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["deep_researcher"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.10",
    "ruff>=0.5",
    "bandit>=1.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
